Copy over test_dpr_prepare_during_year_data.py
