--- tests/unit/test_prepare_during_year_data.py
@@ -1,52 +1,0 @@
-import unittest
-import warnings
-
-from pyspark.sql import SparkSession
-from pyspark.sql.types import (
-    StructType,
-    StructField,
-    StringType,
-    FloatType,
-)
-
-import utils.direct_payments_utils.prepare_direct_payments.prepare_during_year_data as job
-from utils.direct_payments_utils.direct_payments_column_names import (
-    DirectPaymentColumnNames as DP,
-)
-
-
-class TestPrepareDuringYearData(unittest.TestCase):
-    def setUp(self):
-        self.spark = SparkSession.builder.appName(
-            "test_areas_including_carers"
-        ).getOrCreate()
-
-        warnings.simplefilter("ignore", ResourceWarning)
-
-    def test_calculate_total_dprs_during_year_returns_correct_sum(
-        self,
-    ):
-        rows = [
-            ("area_1", 100.0, 2.5),
-            ("area_2", 25.0, 2.5),
-        ]
-        test_schema = StructType(
-            [
-                StructField(DP.LA_AREA, StringType(), False),
-                StructField(DP.SERVICE_USER_DPRS_DURING_YEAR, FloatType(), True),
-                StructField(DP.CARER_DPRS_DURING_YEAR, FloatType(), True),
-            ]
-        )
-        df = self.spark.createDataFrame(rows, schema=test_schema)
-        output_df = job.calculate_total_dprs_during_year(df)
-
-        output_df_list = output_df.sort(DP.LA_AREA).collect()
-
-        self.assertEqual(
-            output_df_list[0][DP.TOTAL_DPRS_DURING_YEAR],
-            102.5,
-        )
-        self.assertEqual(
-            output_df_list[1][DP.TOTAL_DPRS_DURING_YEAR],
-            27.5,
-        )
